backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
alembic>=1.10.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=1.10.7
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
tenacity>=8.2.2
